generalized SpriteManagers' request function to vanilla to work with any byte array. BasicRenderPipeline now has a bitmore functionality that makes it easier to create quick shader programs to render things. RenderProgram now uses BasicRenderPipline to prevent repeated code

+++ shaders/vertex/vertexShader.h
@@ -2,7 +2,7 @@
 
 layout (location = 0) in vec4 values;
 layout (location = 1) in mat4 transform;
-layout (location = 5) in float depth;
+layout (location = 5) in int depth;
 layout (location = 6) in float effect;
